separate test helpers by role
