+++ examples/ping/tests/blockdaemon.spec.ts
@@ -1,11 +1,7 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-    test,
-    expect,
-    WalletGateway,
-} from '@canton-network/core-wallet-test-utils'
+import { test, WalletGateway } from '@canton-network/core-wallet-test-utils'
 import { Page } from '@playwright/test'
 import {
     clickCreatePingContract,
@@ -15,50 +11,8 @@ import {
     expectTxStatusInDappEvents,
     allocateExternalSigningParty,
     createPingContractAndApproveExternal,
-    toMockEndpoint,
-    isLocalhost,
-} from './external-signing-test-helpers.js'
-
-const blockdaemonApiUrl = process.env.BLOCKDAEMON_API_URL
-
-async function setMockBlockdaemonTransactionState(
-    txId: string,
-    status: 'signed' | 'rejected' | 'failed'
-): Promise<void> {
-    const isMockedApi =
-        blockdaemonApiUrl && isLocalhost(new URL(blockdaemonApiUrl))
-    if (!isMockedApi) {
-        return
-    }
-    const promoteResponse = await fetch(
-        toMockEndpoint(blockdaemonApiUrl, '/_admin/setTransactionState'),
-        {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({
-                txId,
-                status,
-            }),
-        }
-    )
-    expect(promoteResponse.ok).toBeTruthy()
-
-    const txResponse = await fetch(
-        toMockEndpoint(blockdaemonApiUrl, '/getTransaction'),
-        {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({ txId }),
-        }
-    )
-    expect(txResponse.ok).toBeTruthy()
-    const tx = (await txResponse.json()) as {
-        txId: string
-        status: string
-    }
-    expect(tx.txId).toBe(txId)
-    expect(tx.status).toBe(status)
-}
+} from './ping-test-helpers.js'
+import { setMockBlockdaemonTransactionState } from './external-signing-test-helpers.js'
 
 test.describe('Blockdaemon external signing', () => {
     test.describe.configure({ mode: 'serial' })
